+++ lib-extra/src/test/java/com/diffplug/spotless/TestP2Provisioner.java
@@ -118,10 +118,12 @@ public class TestP2Provisioner {
 
 		@Override
 		public boolean equals(Object o) {
-			if (this == o)
+			if (this == o) {
 				return true;
-			if (o == null || getClass() != o.getClass())
+			}
+			if (o == null || getClass() != o.getClass()) {
 				return false;
+			}
 			CacheKey cacheKey = (CacheKey) o;
 			return useMavenCentral == cacheKey.useMavenCentral &&
 					Objects.equals(p2Repos, cacheKey.p2Repos) &&