+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/kotlin/Kotlin.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2023 DiffPlug
+ * Copyright 2016-2026 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -48,5 +48,9 @@ public class Kotlin extends FormatterFactory {
 
 	public void addDiktat(Diktat diktat) {
 		addStepFactory(diktat);
+	}
+
+	public void addTableTestFormatter(TableTestFormatter tableTestFormatter) {
+		addStepFactory(tableTestFormatter);
 	}
 }